QTreeView - clean up comments (do not add spans)
Maybe spans would be better in QTreeView than QHeaderView - but
from what I have seen of them (complex and slow) I doubt that
it is what we want in QTreeView.

Beside that the code removes a dead line of code, that is bug.
This is a comment fix that does not change the code.

Change-Id: I4c67a95c4bd2904612d6146077b7cbd59de7bdd6
Reviewed-by: Stephen Kelly <stephen.kelly@kdab.com>