Fixed crash associated with reparenting a QGLWidget
We need to make sure that we don't reset() the context when the
QGLWidget is being reparented, as that will lead to the QOpenGLContext
being destroyed and not recreated, leading to a crash in makeCurrent().

Also, don't destroy() the widget if it has not yet been created, as in
that case there's no ParentAboutToChange event sent.

Task-number: QTBUG-31016

Change-Id: I409fff94456802a80bd72b470a6fbaee87505baa
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Andy Nichols <andy.nichols@digia.com>