Don't depend on the command for extra compilers automatically
For vcproj files we should not automatically add a dependency for extra
compilers as this should be done via the depends variable for the extra
compiler instead.

This fixes a problem where something like '@echo command' was used in the
command as it would depend on echo and not the actual command anyway.

Change-Id: I7c0aa0d62143aa39c518e8bce65f302afdba33c9
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>