Implement system locale for Android
On Android, we would default to C locale always. We need a connection
to Java in order to get the locale, so this has to be done in the
platform plugin rather than the QtCore library. It's enough to
instantiate the QSystemLocale subclass, since this will automatically
register itself as the current system locale.

Task-number: QTBUG-31651
Change-Id: I76f3e30f7dff90e8101cb560cee2b96c9300d9af
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>